move class into stable package
